promote MinIo to test suite
