+++ .github/workflows/regression_suite.yaml
@@ -23,8 +23,21 @@ jobs:
       - name: install optional requirements
         run:  pip install -r $GITHUB_WORKSPACE/requirements-optional.txt
 
+      - name: Setup minio
+        run: |
+          docker run -d -p 9000:9000 --name minio \
+                     -e "MINIO_ACCESS_KEY=minioadmin" \
+                     -e "MINIO_SECRET_KEY=minioadmin" \
+                     -v /tmp/data:/data \
+                     -v /tmp/config:/root/.minio \
+                     minio/minio server /data
+
       - name: run regression tests
         run: coverage run -m pytest 
+        env:
+          MINIO_END_POINT=127.0.0.1:9000
+          MINIO_ACCESS_KEY=minioadmin
+          MINIO_SECRET_KEY=minioadmin
 
       - name: report on coverage
         run: coverage report --include=mabel* --fail-under=80
